fix more stuff
